// client/src/app/models/enums/game-action.enum.ts
export enum GameAction {
    TakeDeckCard = 'takeDeckCard',
    TakeHandCard = 'takeHandCard',
    PlayTableCard = 'playTableCard',
    PlayHandCard = 'playHandCard',
    DropTableCard = 'dropTableCard',
    DropHandCard = 'dropHandCard',
    GiveHandCard = 'giveHandCard',
    GiveTableCard = 'giveTableCard',
    ShowCards = 'showCards',
    SkipShowingCards = 'skipShowingCards',
    AcceptRequest = 'acceptRequest'
}